Renamed & moved BundleTypePrintInfo and related classes
