+++ src/dotnet-core-uninstall/MacOs/SupportedBundleTypeConfigs.cs
@@ -4,11 +4,11 @@ using System.CommandLine.Rendering.Views;
 using System.Linq;
 using Microsoft.DotNet.Tools.Uninstall.Shared.BundleInfo;
 using Microsoft.DotNet.Tools.Uninstall.Shared.BundleInfo.Versioning;
-using Microsoft.DotNet.Tools.Uninstall.Shared.Commands;
+using Microsoft.DotNet.Tools.Uninstall.Shared.Configs;
 
 namespace Microsoft.DotNet.Tools.Uninstall.MacOs
 {
-    internal static class ListCommandExec
+    internal static class SupportedBundleTypeConfigs
     {
         private static readonly Func<IList<Bundle>, GridView> _gridViewGeneratorWithArch = bundles =>
         {
